pyproject.toml
[tool.poetry]
name = "swipematch"
version = "0.1.0"
description = ""
authors = ["RocketMenace <grizzlygit777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.11"}
asyncpg = "^0.30.0"
uvloop = "^0.21.0"
httpx = "^0.28.1"
httptools = "^0.6.4"
pytest = "^8.3.5"
pydantic-settings = "^2.8.1"
sqlalchemy = "^2.0.38"
pytest-asyncio = "^0.26.0"
rich = "^13.9.4"
alembic = "^1.15.2"
email-validator = "^2.2.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.9"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
